Pin Triton to a specific commit for AMD Llama2 submission
The Dockerfile used for building vLLM on ROCm points to the Triton
main branch by default.  This results in a build that is not repeatable,
and recent Triton updates have introduced incompatibilities which cause
the build to fail.

Update the build_llama2.sh script to build vLLM with a specific commit
of Triton; the revision used here is the same revision that was used
for the MLPerf 4.1 submission.

Also make the second stage of the build (adding the MLPerf-specific
code on top of the generic vLLM image) conditional on having a
successful vLLM build.  Without this change, vLLM build failures
would result in a confusing error message.