AMD: Remove call to package submission tool
The run_scenarios.sh script used in AMD's Llama-2 submission includes
a call to a submission packaging tool, but this tool was not part of
the submission package.  While AMD looks forward to including this
tool in a future submission to make it easier for others to submit
MLPerf results with AMD GPUs, the packaging tool is not yet ready
for broader use.  We are removing the call from the run_scenarios.sh
script to eliminate an error message (which doesn't affect the actual
runs).